Delete duplicate schema_to_json_string and export_schema functions, move ad-hoc functions to other.py
